// src/modules/knowledge/enums/knowledge-node-kind.enum.ts
export enum KnowledgeNodeKind {
  ArchitecturalComponent = 'architectural_component',
  DomainConcept = 'domain_concept',
  BusinessRule = 'business_rule',
  Workflow = 'workflow',
  WorkflowStep = 'workflow_step',
  State = 'state',
  StateTransition = 'state_transition',
  DomainEvent = 'domain_event',
  EventHandler = 'event_handler',
}